+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/network/users/EditUserRequest.java
@@ -7,13 +7,15 @@ import com.android.volley.Response;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import org.json.JSONException;
 import org.json.JSONObject;
+
+import java.util.HashMap;
 
 import blueprint.com.sage.models.APIError;
 import blueprint.com.sage.models.User;
 import blueprint.com.sage.network.BaseRequest;
 import blueprint.com.sage.utility.network.NetworkManager;
-import blueprint.com.sage.utility.network.NetworkUtils;
 
 /**
  * Created by charlesx on 11/24/15.
@@ -22,7 +24,7 @@ public class EditUserRequest extends BaseRequest {
     public EditUserRequest(final Activity activity, User user,
                            final Response.Listener<User> onSuccess,
                            final Response.Listener<APIError> onFailure) {
-        super(Method.PUT, makeUrl(null, "users", String.valueOf(user.getId())), NetworkUtils.convertToUserParams(user),
+        super(Method.PUT, makeUrl(null, "users", String.valueOf(user.getId())), convertToUserParams(user),
                 new Response.Listener<JSONObject>() {
                     @Override
                     public void onResponse(JSONObject o) {
@@ -41,5 +43,30 @@ public class EditUserRequest extends BaseRequest {
                         onFailure.onResponse(error);
                     }
                 }, activity);
+    }
+
+    public static JSONObject convertToUserParams(User user) {
+        HashMap<String, JSONObject> params = new HashMap<>();
+        JSONObject userObject = new JSONObject();
+
+        try {
+            userObject.put("email", user.getEmail());
+            userObject.put("first_name", user.getFirstName());
+            userObject.put("last_name", user.getLastName());
+            if (user.getPassword() != null) {
+                userObject.put("password", user.getPassword());
+                userObject.put("confirm_password", user.getConfirmPassword());
+            }
+            userObject.put("school_id", user.getSchoolId());
+            userObject.put("current_password", user.getCurrentPassword());
+            if (user.getProfileData() != null) {
+                userObject.put("data", user.getProfileData());
+            }
+        } catch(JSONException e) {
+            Log.e(CreateUserRequest.class.toString(), e.toString());
+        }
+
+        params.put("user", userObject);
+        return new JSONObject(params);
     }
 }
